Add inplace for Vec3

+++ nept.core/src/main/java/app/neptune/NeptVec.java
@@ -13,9 +13,9 @@ public final class NeptVec {
     public static void call() {
         try (Arena arena = Arena.ofConfined()) {
             MemorySegment vec = arena.allocate(Vec3Layouts.VEC3);
-            Vec3Layouts.X.set(vec, 0, 3f);
-            Vec3Layouts.Y.set(vec, 0, 4f);
-            Vec3Layouts.Z.set(vec, 0, 0f);
+            Vec3Layouts.X.set(vec, (long) 0, 3f);
+            Vec3Layouts.Y.set(vec, (long) 0, 4f);
+            Vec3Layouts.Z.set(vec, (long) 0, 0f);
 
             float len = (float) Vec3Layouts.VEC3_LENGTH.invokeExact(vec);
             IO.println("vec3_length");
@@ -24,23 +24,23 @@ public final class NeptVec {
             Vec3Layouts.VEC3_NORMALIZE.invokeExact(vec);
             IO.println("vec3_normalize");
             IO.println(
-                    Vec3Layouts.X.get(vec, 0) + ", " +
-                    Vec3Layouts.Y.get(vec, 0) + ", " +
-                    Vec3Layouts.Z.get(vec, 0)
+                    (float) Vec3Layouts.X.get(vec, (long) 0) + ", " +
+                    (float) Vec3Layouts.Y.get(vec, (long) 0) + ", " +
+                    (float) Vec3Layouts.Z.get(vec, (long) 0)
             );
 
             MemorySegment vec2 = arena.allocate(Vec3Layouts.VEC3);
-            Vec3Layouts.X.set(vec2, 0, 2f);
-            Vec3Layouts.Y.set(vec2, 0, 3f);
-            Vec3Layouts.Z.set(vec2, 0, 1f);
+            Vec3Layouts.X.set(vec2, (long) 0, 2f);
+            Vec3Layouts.Y.set(vec2, (long) 0, 3f);
+            Vec3Layouts.Z.set(vec2, (long) 0, 1f);
 
             var vecOut = (MemorySegment) Vec3Layouts.VEC3_ADD
                     .invokeExact((SegmentAllocator) arena, vec, vec2);
             IO.println("vec3_add");
             IO.println(
-                    Vec3Layouts.X.get(vecOut, 0) + ", " +
-                    Vec3Layouts.Y.get(vecOut, 0) + ", " +
-                    Vec3Layouts.Z.get(vecOut, 0)
+                    (float) Vec3Layouts.X.get(vecOut, (long) 0) + ", " +
+                    (float) Vec3Layouts.Y.get(vecOut, (long) 0) + ", " +
+                    (float) Vec3Layouts.Z.get(vecOut, (long) 0)
             );
         } catch (Throwable e) {
             throw new RuntimeException(e);
@@ -54,16 +54,16 @@ public final class NeptVec {
         if (a.isEmpty()) return;
 
         try (Arena arena = Arena.ofConfined()) {
-            int count = a.size();
+            long count = a.size();
 
             MemorySegment segA = arena.allocate(Vec3Layouts.VEC3, count);
             MemorySegment segB = arena.allocate(Vec3Layouts.VEC3, count);
             MemorySegment segR = arena.allocate(Vec3Layouts.VEC3, count);
 
             long stride = Vec3Layouts.VEC3.byteSize();
-            for (int offset = 0; offset < count; offset++) {
-                Vec3 va = a.get(offset);
-                Vec3 vb = b.get(offset);
+            for (long offset = 0L; offset < count; offset++) {
+                Vec3 va = a.get((int) offset);
+                Vec3 vb = b.get((int) offset);
 
                 Vec3Layouts.X.set(segA, stride * offset, va.x());
                 Vec3Layouts.Y.set(segA, stride * offset, va.y());
@@ -74,7 +74,7 @@ public final class NeptVec {
                 Vec3Layouts.Z.set(segB, stride * offset, vb.z());
             }
 
-            Vec3Layouts.VEC3_ADD_ALL.invokeExact(segA, segB, segR, count);
+            Vec3Layouts.VEC3_ADD_ALL.invokeExact(segA, segB, segR, (int) count);
         } catch (Throwable e) {
             throw new RuntimeException(e);
         }
